Improve inline rendering performance
Inline rendering doesn't require load the full controller and simulate a
request. Instead, the controller for async loads and the helper for
inline loads, use the `Display` class that all the components controller
should implement.

Only one `get` action is available per controller, by default it is the
`display` action.